django/project/interface/digitalsignage/requirements.txt
opencv_contrib_python==4.1.1.26
scipy==1.3.3
matplotlib==3.1.1
dlib==19.18.0
tensorflow_gpu==1.14.0
numpy==1.17.2
scikit_learn==0.22
